+++ packages/documentation/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tidbcloud/uikit-documentation
 
+## 0.1.30-beta.72
+
+### Patch Changes
+
+- fix ProTable footer style
+- Updated dependencies
+  - @tidbcloud/uikit@2.0.0-beta.72
+
 ## 0.1.30-beta.71
 
 ### Patch Changes